refactor new useInitializeFlowIntegration hook
